chore(asm): refactor
