+++ ddtrace/appsec/iast/_taint_tracking/__init__.py
@@ -6,4 +6,5 @@ from ddtrace.appsec._asm_request_context import get_tainted_ranges  # noqa: F401
 from ddtrace.appsec._asm_request_context import is_pyobject_tainted  # noqa: F401
 from ddtrace.appsec._asm_request_context import set_tainted_ranges  # noqa: F401
 from ddtrace.appsec._asm_request_context import taint_pyobject  # noqa: F401
-from ddtrace.appsec.iast._taint_tracking import setup  # type: ignore[attr-defined]  # noqa: F401
+from ddtrace.appsec.iast._taint_tracking._native import new_pyobject_id  # noqa: F401
+from ddtrace.appsec.iast._taint_tracking._native import setup  # noqa: F401
