+++ ddtrace/bootstrap/sitecustomize.py
@@ -163,7 +163,7 @@ try:
 
             from ddtrace.appsec.iast._ast.ast_patching import _should_iast_patch
             from ddtrace.appsec.iast._loader import _exec_iast_patched_module
-            from ddtrace.appsec.iast._taint_tracking import setup  # type:ignore[attr-defined]
+            from ddtrace.appsec.iast._taint_tracking import setup
             from ddtrace.internal.module import ModuleWatchdog
 
             setup(bytes.join, bytearray.join)
