+++ tests/appsec/iast/aspects/test_encode_decode_aspect.py
@@ -32,10 +32,10 @@ def catch_all(fun, args, kwargs):
 @pytest.mark.skipif(sys.version_info < (3, 6, 0), reason="Python 3.6+ only")
 def test_decode_and_add_aspect(infix, args, kwargs, should_be_tainted, prefix, suffix):
     import ddtrace.appsec.iast._ast.aspects as ddtrace_aspects
-    from ddtrace.appsec.iast._new_taint_tracking import clear_taint_mapping
-    from ddtrace.appsec.iast._new_taint_tracking import get_tainted_ranges
-    from ddtrace.appsec.iast._new_taint_tracking import setup
-    from ddtrace.appsec.iast._new_taint_tracking import taint_pyobject
+    from ddtrace.appsec.iast._taint_tracking import clear_taint_mapping
+    from ddtrace.appsec.iast._taint_tracking import get_tainted_ranges
+    from ddtrace.appsec.iast._taint_tracking import setup
+    from ddtrace.appsec.iast._taint_tracking import taint_pyobject
 
     setup(bytes.join, bytearray.join)
     clear_taint_mapping()
@@ -77,10 +77,10 @@ def test_decode_and_add_aspect(infix, args, kwargs, should_be_tainted, prefix, s
 @pytest.mark.skipif(sys.version_info < (3, 6, 0), reason="Python 3.6+ only")
 def test_encode_and_add_aspect(infix, args, kwargs, should_be_tainted, prefix, suffix):
     import ddtrace.appsec.iast._ast.aspects as ddtrace_aspects
-    from ddtrace.appsec.iast._new_taint_tracking import clear_taint_mapping
-    from ddtrace.appsec.iast._new_taint_tracking import get_tainted_ranges
-    from ddtrace.appsec.iast._new_taint_tracking import setup
-    from ddtrace.appsec.iast._new_taint_tracking import taint_pyobject
+    from ddtrace.appsec.iast._taint_tracking import clear_taint_mapping
+    from ddtrace.appsec.iast._taint_tracking import get_tainted_ranges
+    from ddtrace.appsec.iast._taint_tracking import setup
+    from ddtrace.appsec.iast._taint_tracking import taint_pyobject
 
     setup(bytes.join, bytearray.join)
     clear_taint_mapping()
