+++ tests/appsec/iast/aspects/test_str_aspect.py
@@ -48,10 +48,10 @@ def test_str_aspect(obj, kwargs):
 @pytest.mark.skipif(sys.version_info < (3, 6, 0), reason="Python 3.6+ only")
 def test_str_aspect_tainting(obj, kwargs, should_be_tainted):
     import ddtrace.appsec.iast._ast.aspects as ddtrace_aspects
-    from ddtrace.appsec.iast._new_taint_tracking import clear_taint_mapping
-    from ddtrace.appsec.iast._new_taint_tracking import is_pyobject_tainted
-    from ddtrace.appsec.iast._new_taint_tracking import setup
-    from ddtrace.appsec.iast._new_taint_tracking import taint_pyobject
+    from ddtrace.appsec.iast._taint_tracking import clear_taint_mapping
+    from ddtrace.appsec.iast._taint_tracking import is_pyobject_tainted
+    from ddtrace.appsec.iast._taint_tracking import setup
+    from ddtrace.appsec.iast._taint_tracking import taint_pyobject
 
     setup(bytes.join, bytearray.join)
     clear_taint_mapping()
