+++ tests/appsec/iast/test_taint_utils.py
@@ -5,9 +5,9 @@ import pytest
 
 try:
     from ddtrace.appsec.iast._input_info import Input_info
-    from ddtrace.appsec.iast._new_taint_tracking import is_pyobject_tainted
-    from ddtrace.appsec.iast._new_taint_tracking import setup as taint_tracking_setup
-    from ddtrace.appsec.iast._new_taint_tracking import taint_pyobject
+    from ddtrace.appsec.iast._taint_tracking import is_pyobject_tainted
+    from ddtrace.appsec.iast._taint_tracking import setup as taint_tracking_setup
+    from ddtrace.appsec.iast._taint_tracking import taint_pyobject
     from ddtrace.appsec.iast._taint_utils import LazyTaintDict
     from ddtrace.appsec.iast._taint_utils import check_tainted_args
 except (ImportError, AttributeError):
