+++ tests/contrib/django/test_django_appsec.py
@@ -722,8 +722,8 @@ def test_request_suspicious_request_block_match_response_headers(client, test_sp
 
 @pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
 def test_django_tainted_user_agent_iast_enabled(client, test_spans, tracer):
-    from ddtrace.appsec.iast._new_taint_tracking import clear_taint_mapping
-    from ddtrace.appsec.iast._new_taint_tracking import setup
+    from ddtrace.appsec.iast._taint_tracking import clear_taint_mapping
+    from ddtrace.appsec.iast._taint_tracking import setup
 
     with override_global_config(dict(_iast_enabled=True)):
         tracer._iast_enabled = True
@@ -746,8 +746,8 @@ def test_django_tainted_user_agent_iast_enabled(client, test_spans, tracer):
 
 @pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
 def test_django_tainted_user_agent_iast_disabled(client, test_spans, tracer):
-    from ddtrace.appsec.iast._new_taint_tracking import clear_taint_mapping
-    from ddtrace.appsec.iast._new_taint_tracking import setup
+    from ddtrace.appsec.iast._taint_tracking import clear_taint_mapping
+    from ddtrace.appsec.iast._taint_tracking import setup
 
     with override_global_config(dict(_iast_enabled=False)):
         tracer._iast_enabled = False
